Fix sorting of links and contacts
In the final version of reports 2019q1 and 2019q2 we always had links
before contacts, however the templates and the reports in the repository
have contacts before links.

This commit fixes the inconsistency for 2019q3 and templates by putting
links before contacts. The inconsistency is still present for older reports.