Add tools to prepare the mail version of reports
The tools consist in a few more targets added to tools/Makefile for
downloading the html version of the report from the freebsd.org website,
converting it and making some prettifications.

A small modification has been made to 2019q4 too: it has been added a
variable needed by the prettify.perl script.