tools/sendcalls: Add support for an extra header
Mails sent to quarterly-calls@ need to be sent with a special header to
pass moderation.
The approach consisting in sending a mail to quarterly-calls@ and CCing
the other destinataries is replaced by sending one single mail per
destinatary: it is thus possible to add the special header to
quarterly-calls@ only.

Update the test output consistently with the above changes (also explictly
closing a file that was left open).