Revert "tools/sendcalls: Add support for an extra header"
This reverts commit 57194337976dbada6168abb1065f0a3ed98464da.
It only maintains the explicit closure of a file that was forgotten open.

I contacted postmaster@ to have my mail address whitelisted for moderation
on quarterly-calls@, so there is no need anymore for extra headers and it
is again possible for me to send the calls to quarterly-calls@ while CCing
the other destinataries.

In the future anyone who needs to send calls through this script without
explicit moderation approval on quarterly-calls@ should probably ask to
have his/her address whitelisted as well (alternatives are manual
moderation, restore of the now reverted commit or similar modification).