tools/md2docbook: Use stricter link regex
Previously, "([foo](http://bar))" would be parsed as an open paren
followed by a link to "http://bar)" with text "foo". Instead, ensure
that the matching square bracket or parenthesis is the first we see to
avoid any such issues.

This is stricter than specifications like CommonMark, but we are already
far more restrictive in what we accept, and it's better to not generate
anchors at all than generate broken ones. In general people also tend
not to use the weirder parts of Markdown.