tools/sendcalls: Add a filter to reduce BCC
As some people receive too many copies of the calls for quarterly
reports, we add a filter to the script so that no address that ends in
'@FreeBSD.org' (case insensitive) is added to the BCC field.

The BCC addresses are taken from the submitters of the precedent
reports. When those addresses end in '@FreeBSD.org', they should
correspond to developers or teams of developers, but all developers
already receive the calls through developers@.