tools/Makefile.template:  Add file
Makefile.template is a template that can be used to create locally the
Makefiles in the quarter directories. It is already initialized with
some standard reports that are expected to be submitted each quarter:
it should always be ensured that all standard reports are included in
the template and sorted correctly.