lsof.adoc: correction and suggestions
The manual page for lsof is in section 8, not 1. 

The + for a hard line break was superfluous, where the next line was blank. 

<https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=266881> aims to raise awareness of macros such as package and man. 

Whilst the Link URL within this report is not part of a long sentence, the markup in this pull request might be good in the context of <https://github.com/freebsd/freebsd-quarterly/issues/530>. 

Whilst here, a handful of suggested changes.